Revert "Merge pull request #64 from Team-EverFlow/feature/53-oauth2-button-feature"
This reverts commit b4017072d7a29e44462eab8452d8e1061a7867c7, reversing
changes made to c4436fc5ca4c2ea8978b5b39a93cd8cdaab8f212.
