+++ src/Router.js
@@ -2,12 +2,6 @@
 
 import React from 'react';
 import { BrowserRouter, Routes, Route } from 'react-router-dom';
-
-import WordTestList from './pages/wordTest/WordTestList';
-import TestPageGoToTest from './pages/wordTest/TestPageGoToTest';
-import TestPageLearnWords from './pages/wordTest/TestPageLearnWords';
-import Login from './pages/Login';
-
 import MainView from './pages/MyBadgesSetting';
 import Profile from './pages/Profile';
 import MyBadges from './pages/MyBadges';
@@ -17,16 +11,6 @@ const Router = () => {
     return (
         <BrowserRouter>
             <Routes>
-                <Route path="/login" element={<Login />} />
-                <Route path="/" element={<WordTestList />} />
-                <Route
-                    path="/TestPageGoToTest"
-                    element={<TestPageGoToTest />}
-                />
-                <Route
-                    path="/TestPageLearnWords"
-                    element={<TestPageLearnWords />}
-                />
                 <Route path="/" element={<MainView />} />
                 <Route path="/profile" element={<Profile />} />
                 <Route path="/badge" element={<MyBadges />} />
